agent/requirements.txt
pystray>=0.19
Pillow>=10.0
watchdog>=4.0
httpx>=0.27
semver>=3.0
pyinstaller>=6.0
tomli>=2.0; python_version < "3.11"
